refactor(dashboard-sync-feedback): extract local sync feedback expiry logic and cleanup related code
